+++ RockLib.Configuration.ObjectFactory/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+#### Changed
+
+- Fixes ambiguous constructor issue between injectable and named parameters counts.
+
 ## 1.6.9 - 2021-08-11
 
 #### Changed